[CI] cleanup single/multi-card test
