+++ tests/e2e/multicard/2-cards/test_offline_weight_load.py
@@ -34,7 +34,7 @@ MODELS = ["Qwen/Qwen3-30B-A3B"]
 def test_qwen3_offline_load_and_sleepmode_tp2(model):
     script = Path(
         __file__
-    ).parent.parent.parent.parent / "examples" / "offline_external_launcher.py"
+    ).parent.parent.parent.parent.parent / "examples" / "offline_external_launcher.py"
     env = os.environ.copy()
     cmd = [
         sys.executable,
