+++ tests/e2e/multicard/4-cards/spec_decode/test_mtp_qwen3_next.py
